requirements.txt
# Core Python dependencies  
python-dotenv==1.1.0    # For managing environment variables  
requests==2.32.3        # For API calls (e.g., weather updates)  

# Speech Recognition  
SpeechRecognition==3.14.2	# For converting speech to text  
pyaudio==0.2.14			# Required for live audio processing  

# Text-to-Speech (TTS)  
pyttsx3==2.98            # Offline TTS  

openai # For OpenAI API access (e.g., GPT-4)

rapidfuzz==3.13.0        # For fuzzy string matching (e.g., command recognition)

# Automation (for launching applications & searching files)  
#pyautogui==0.9.54        # For controlling the OS  
#pygetwindow==0.0.9       # For managing windows  
#keyboard==0.13.5         # For handling keyboard input  

# GUI (Optional, if building a graphical interface)  
PyQt5

# Testing  
#pytest==8.0.0           # For writing unit tests  
